requirements.txt
Flask==0.10.1

Frozen-Flask==0.11

xlwt
xlrd
json
request
urllib3
urllib